create routes home and game file structure
