+++ src/routes/Game/index.js
@@ -1,0 +1,8 @@
+const GamePage = () => {
+    return (
+        <div>
+            This is Game Page!
+        </div>
+    )
+}
+export default GamePage;
